Repurposed project-card-skills into related-skills Used related skills to also render task skills on task-card
