+++ app/templates/components/project-card.hbs
@@ -14,6 +14,6 @@
   <p class="description">
     {{project.description}}
   </p>
-  {{project-card-skills skills=projectSkills}}
+  {{related-skills class="project-card__skills" skills=projectSkills}}
   {{project-card-members members=projectOrganizationMembers}}
 </div>
