+++ app/templates/components/related-skills.hbs
